Move icon declaration away from general index.js
