+++ src/components/icon/icons/index.js
@@ -1,0 +1,114 @@
+import Apartment from './apartment';
+import ArrowDown from './arrowDown';
+import ArrowLeft from './arrowLeft';
+import ArrowRight from './arrowRight';
+import ArrowUp from './arrowUp';
+import ArrowUpDown from './arrowUpDown';
+import Balloon from './balloon';
+import Bell from './bell';
+import Calendar from './calendar';
+import Cart from './cart';
+import Checkmark from './checkmark';
+import CheckmarkLarge from './checkmarkLarge';
+import ChevronDown from './chevronDown';
+import ChevronUp from './chevronUp';
+import CircleArrow from './circleArrow';
+import CircleSlash from './circleSlash';
+import Clock from './clock';
+import Close from './close';
+import Cloud from './cloud';
+import Edit from './edit';
+import Ellipsis from './ellipsis';
+import ExclamationPoint from './exclamationPoint';
+import Expand from './expand';
+import Film from './film';
+import Filter from './filter';
+import FloppyDisk from './floppyDisk';
+import Folder from './folder';
+import GraphArea from './graphArea';
+import GraphCandlestick from './graphCandlestick';
+import GraphLine from './graphLine';
+import GraphVolume from './graphVolume';
+import Heart from './heart';
+import Key from './key';
+import LightningBolt from './lightningBolt';
+import Lock from './lock';
+import Mail from './mail';
+import MailOpen from './mailOpen';
+import News from './news';
+import OrderDepth from './orderDepth';
+import Print from './print';
+import Profile from './profile';
+import Questionmark from './questionmark';
+import SearchIcon from './search';
+import Share from './share';
+import SignOut from './signOut';
+import SocialFacebook from './socialFacebook';
+import SocialGplus from './socialGplus';
+import SocialInstagram from './socialInstagram';
+import SocialTwitter from './socialTwitter';
+import Star from './star';
+import Tag from './tag';
+import TickingClock from './tickingClock';
+import Trash from './trash';
+import VerticalEllipsis from './verticalEllipsis';
+
+const Icon = {
+  Apartment,
+  ArrowDown,
+  ArrowLeft,
+  ArrowRight,
+  ArrowUp,
+  ArrowUpDown,
+  Balloon,
+  Bell,
+  Calendar,
+  Cart,
+  Checkmark,
+  CheckmarkLarge,
+  ChevronDown,
+  ChevronUp,
+  CircleArrow,
+  CircleSlash,
+  Clock,
+  Close,
+  Cloud,
+  Edit,
+  Ellipsis,
+  ExclamationPoint,
+  Expand,
+  Film,
+  Filter,
+  FloppyDisk,
+  Folder,
+  GraphArea,
+  GraphCandlestick,
+  GraphLine,
+  GraphVolume,
+  Heart,
+  Key,
+  LightningBolt,
+  Lock,
+  Mail,
+  MailOpen,
+  News,
+  OrderDepth,
+  Print,
+  Profile,
+  Questionmark,
+  Search: SearchIcon,
+  Share,
+  SignOut,
+  SocialFacebook,
+  SocialGplus,
+  SocialInstagram,
+  SocialTwitter,
+  Star,
+  Tag,
+  TickingClock,
+  Trash,
+  VerticalEllipsis,
+};
+
+export default Icon;
+
